requirements.txt
pandas
googlesearch-python
beautifulsoup4
requests
google-search-results
python-dotenv
google-genai
streamlit
